+++ README.md
@@ -8,6 +8,7 @@ README.md - этот файл
 requirements.txt - основные зависимости
 ./conftest.py - конфигурация тестов
 ./locators.py - локаторы
+./data.py - ответы на вопросы вниза страницы
 
 ./tests/ - каталог с тестами
 ./tests/test_order_page.py - тесты для проверки создания заказа
@@ -16,8 +17,8 @@ requirements.txt - основные зависимости
 
 ./pages/ - каталог с классами страниц
 ./pages/order_page.py - класс для страницы оформления заказа
-./pages/link_page.py - класс для базовой страницы
+./pages/base_page.py - класс для базовой страницы
 ./pages/questions_page.py - класс для страницы с вопросами
 
 ./allure_results/ - каталог с Allure отчетом
-```# Sprint_4_
+```
